requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.4.0
pillow==10.2.0
requests==2.31.0
numpy==1.26.3
python-dotenv
pytest
httpx